+++ R/process-data.R
@@ -25,94 +25,7 @@
 #' }
 #'
 tsg_process_data <- function(x, min_coverage = 0, verbose = TRUE) {
-  for (i in seq_along(x)) {
-    if (any(x[[i]][["data_type"]] == "json")) {
-      x[[i]] <- tidyr::unnest_wider(x[[i]], "funding_organization",
-        names_sep = "_"
-      )
-      x[[i]] <- tidyr::unnest_wider(x[[i]], "recipient_organization",
-        names_sep = "_"
-      )
-
-      if ("recipient_organization_location" %in% colnames(x[[i]])) {
-        x[[i]][["recipient_organization_location"]] <- purrr::flatten(
-          x[[i]][["recipient_organization_location"]]
-        )
-        x[[i]] <- tidyr::unnest_wider(x[[i]], "recipient_organization_location",
-          names_sep = "_"
-        )
-      }
-
-      if ("funding_organization_location" %in% colnames(x[[i]])) {
-        x[[i]][["funding_organization_location"]] <- purrr::flatten(
-          x[[i]][["funding_organization_location"]]
-        )
-        x[[i]] <- tidyr::unnest_wider(x[[i]], "funding_organization_location",
-          names_sep = "_"
-        )
-      }
-
-      if ("planned_dates" %in% colnames(x[[i]])) {
-        x[[i]] <- tidyr::unnest_wider(x[[i]], "planned_dates", names_sep = "_")
-      }
-
-      if ("grant_programme" %in% colnames(x[[i]])) {
-        x[[i]] <- tidyr::unnest_wider(x[[i]], "grant_programme",
-          names_sep = "_"
-        )
-      }
-
-      if ("beneficiary_location" %in% colnames(x[[i]])) {
-        x[[i]] <- tidyr::unnest_wider(x[[i]], "beneficiary_location",
-          names_sep = "_"
-        )
-      }
-
-      if ("classifications" %in% colnames(x[[i]])) {
-        x[[i]] <- tidyr::unnest_wider(x[[i]], "classifications",
-          names_sep = "_"
-        )
-
-        if ("classifications_code" %in% colnames(x[[i]])) {
-          x[[i]][["classifications_code"]] <- unlist(lapply(
-            x[[i]][["classifications_code"]], paste,
-            collapse = ", "
-          ))
-        }
-
-        if ("classifications_title" %in% colnames(x[[i]])) {
-          x[[i]][["classifications_title"]] <- unlist(lapply(
-            x[[i]][["classifications_title"]], paste,
-            collapse = ", "
-          ))
-        }
-
-        if ("classifications_url" %in% colnames(x[[i]])) {
-          x[[i]][["classifications_url"]] <- unlist(lapply(
-            x[[i]][["classifications_url"]], paste,
-            collapse = ", "
-          ))
-        }
-
-        if ("classifications_vocabulary" %in% colnames(x[[i]])) {
-          x[[i]][["classifications_vocabulary"]] <- unlist(lapply(
-            x[[i]][["classifications_vocabulary"]], paste,
-            collapse = ", "
-          ))
-        }
-      }
-
-      x[[i]] <- janitor::clean_names(x[[i]])
-      x[[i]] <- janitor::remove_empty(x[[i]], which = "cols")
-
-      names(x[[i]]) <- gsub("organization", "org", names(x[[i]]))
-      names(x[[i]]) <- gsub("org_id$", "org_identifier", names(x[[i]]))
-    }
-    if (verbose) message(paste0("Processing ", i, " of ", length(x)))
-  }
-
-  x <- lapply(x, dplyr::mutate_all, as.character)
-  df <- dplyr::bind_rows(x)
+  df <- tsg_core_process(x, verbose, process_type = "all")
 
   if (min_coverage > 0) {
     if (min_coverage > 1 | min_coverage < 0) {
@@ -130,18 +43,6 @@ tsg_process_data <- function(x, min_coverage = 0, verbose = TRUE) {
       df <- df[names(df) %in% min_df$names]
     }
   }
-
-  req_list <- c(
-    "identifier", "title", "description", "currency",
-    "amount_awarded", "award_date", "recipient_org_identifier",
-    "recipient_org_name", "funding_org_identifier",
-    "funding_org_name", "publisher_prefix"
-  )
-
-  df <- dplyr::select(df, req_list, tidyselect::everything())
-
-  df$amount_awarded <- as.numeric(df$amount_awarded)
-  df$award_date <- as.Date(df$award_date)
 
   df
 }