+++ R/utils-req-list.R
@@ -1,0 +1,16 @@
+tsg_req_list <- c(
+  "identifier", "title", "description", "currency",
+  "amount_awarded", "award_date", "recipient_org_identifier",
+  "recipient_org_name", "funding_org_identifier",
+  "funding_org_name", "publisher_prefix"
+)
+
+
+tsg_proc_req_list <-  c(
+  "identifier", "title", "description", "currency",
+  "amount_awarded", "award_date", "recipient_org_identifier",
+  "recipient_org_name", "funding_org_identifier",
+  "funding_org_name", "recipient_org", "funding_org",
+  "recipient_organization", "funding_organization",
+  "publisher_prefix", "license_name", "data_type"
+)